Augment a Data Frame with Custom Color Column

This function adds a new column named 'custom_color' to a given data frame. The new column maps factor levels of a specified column to colors from a user-defined palette.

augment_with_custom_color(
  df,
  column_name,
  color_palette,
  color_palette_name = NULL
)

Arguments

df

A data frame to be augmented.

column_name

The name of the column in df whose factor levels are to be mapped to colors. This column should exist in df and should be a factor or convertible to a factor.

color_palette

A vector of colors, corresponding to the factor levels of the specified column. The number of colors must match the number of unique factor levels in the column.

color_palette_name

optional named color palette

Value

A data frame identical to df but with an additional column 'custom_color', which contains the color mappings.

Examples

data <- data.frame(
  category = c("A", "B", "C", "A", "B"),
  value = c(10, 20, 30, 40, 50)
)
palette <- c("red", "green", "blue")
augmented_data <- augment_with_custom_color(data, "category", palette)