consensusSubtypingPrep

generate input for consensus clustering give several clustering algorithms

consensusSubtypingPrep(
  dataToTrain,
  dataToPredict,
  featureNames,
  clustVec,
  maxK,
  reorderingVariable,
  mvcl,
  idvar,
  visitName,
  baselineVisit,
  whichrank = 0,
  ntoreturnperk = 2,
  verbose = FALSE
)

Arguments

dataToTrain

dataframe input that contains the relevant variables (may have others as well) on which training will be based

dataToPredict

dataframe input that contains the relevant variables (may have others as well) for which prediction will be done

featureNames

names to use in the clustering

clustVec

names of the clustering methods to use

maxK

the maximum desired number of classes

reorderingVariable

the name of the column to use to reorder the cluster names

mvcl

character prefix for the new cluster column names

idvar

variable name for unique subject identifier column

visitName

the column name defining the visit variables

baselineVisit

the string naming the baseline visit

whichrank

allows user to get 2nd (or 3rd) rank set of methods

ntoreturnperk

number of method results per k to return

verbose

boolean

Value

new dataframe with new variables attached

Author

Avants BB

Examples

mydf = generateSubtyperData( 100 )