subtype plotting

Assemble a set of standard plots looking at subtype results to support comparing across a hierarchy of types both cross-sectionally and longitudinally.

hierarchicalSubtypePlots(
  inputDataFrame,
  variableToVisualize,
  hierarchyOfSubtypes,
  idvar,
  vizname,
  whiskervar = c("ci", "se"),
  consistentSubset = FALSE,
  manualColors,
  outputPrefix,
  width = 12,
  height = 8
)

Arguments

inputDataFrame

Input complete data frame

variableToVisualize

string naming the variable to display across subtypes

hierarchyOfSubtypes

string vector of subtypes with increasing degrees of specificity

idvar

variable name for unique subject identifier column

vizname

the name of the grouped time variable (e.g. years change rounded to nearest quarter year)

whiskervar

character either ci or se

consistentSubset

display longitudinal data only from subjects that are consistently present at all visits

manualColors

a list of user defined manual colors; the length of this list should match the length of hierarchyOfSubtypes and colors should be named according to the levels therein. each entry in the list should be a string vector of color names.

outputPrefix

filename prefix for the stored pdf plots; if missing, just plot to display

width

the width of the graphics region in inches.

height

the height of the graphics region in inches.

Value

the output is a set of plots saved at the outputPrefix location

Author

Avants BB

Examples

mydf = generateSubtyperData( 1000 )
qdf = trainSubtypeUni( mydf, "cognition", c("C0","C1","C2"), c(0.33,0.66) )
qdf = predictSubtypeUni( mydf, qdf, "Id" )
if (FALSE) { # \dontrun{
hierarchicalSubtypePlots( qdf, "cognition", c("DX", "subtype" ),
 "Id", "visit", outputPrefix='/tmp/X' )
} # }