Predict subtype for univariate data

This is the inference module for subtype definition based on a vector. After training, we can predict subtype very easily in new data based on the data frame produced by training trainSubtypeUni. If one passes the visitName to the function then we will define subtype from the baseline value alone.

predictSubtypeUni(
  mxdfin,
  subtypeDataFrame,
  idvar,
  visitName,
  baselineVisit,
  rename = TRUE
)

Arguments

mxdfin
Input data frame

subtypeDataFrame
data frame defining the subtypes and cutpoints

idvar
variable name for unique subject identifier column

visitName
the column name defining the visit variables

baselineVisit
the string naming the baseline visit

rename
boolean will rename levels to user-provided names

Value

data frame with attached subtypess

Author

Avants BB

Examples

mydf = generateSubtyperData( 100 )
rbfnames = names(mydf)[grep("Random",names(mydf))]
mydf = outlierness( mydf, rbfnames )

#> [1] "r update: 2"
#> [1] "r update: 3"
#> [1] "r update: 4"
#> [1] "r update: 5"
#> [1] "r update: 6"
#> [1] "r update: 7"
#> [1] "r update: 8"

#> Error in `$<-.data.frame`(`*tmp*`, "OL_KNN_IN", value = c(-5, -3, -18, -5, -11, -8, -6, -11, -3, -3, -9, -5, -9, -6, -20, -18, -6, -1, -9, -9, -6, -20, -10, -10, -9, -15, -4, -9, -15, -19, -2, -1, -7, -9, -2, 0, -7, -19, -6, -13, -2, -3, 0, -9, -3, -11, -3, -7, -7, -14, -9, -11, -14, -12, -12, -1, -15, -9, -11, -3, -8, -5, -5, -7, -3, -4, -7, -10, -6, -8, -4, -1, -5, -11, -3, -12, -11, -5, -4, -12, 0, -14, -5, -12, -6, -12, -10, 0, -5, -23, -10, -7, -5, -15, -1, -9, -6, -12, -13)): replacement has 99 rows, data has 100

mydf = highestQualityRepeat( mydf, "Id", "visit", "OL_KNN_SUM")

#> Error in highestQualityRepeat(mydf, "Id", "visit", "OL_KNN_SUM"): qualityvar not in dataframe

qdf = trainSubtypeUni( mydf, "cognition", c("C0","C1","C2"), c(0.33,0.66) )
pdf = predictSubtypeUni( mydf, qdf, "Id" )