Create a Table 1 for Population Characteristics

This function generates a "Table 1" for summarizing population characteristics in an academic paper. It uses the gtsummary package for summarizing and the gt package for table customization, including adding stripes. Additionally, if the output format is LaTeX, it generates the corresponding LaTeX code.

create_table1(
  data,
  summary_vars,
  group_var,
  title = "Table 1: Population Characteristics",
  output_format = "gt",
  landscape = TRUE,
  latex_font_size = c(`12.0pt` = "8.0pt", `14.4pt` = "10.0pt")
)

Arguments

data

A data frame containing the dataset.

summary_vars

A character vector of variable names to include in the summary.

group_var

A character string of the variable name to group by.

title

A character string for the table title. Default is "Table 1: Population Characteristics".

output_format

A character string indicating the desired output format, either "gt" or "latex". Default is "gt".

landscape

boolean

latex_font_size

optional font size changes

Value

A gt table object or LaTeX code depending on the output_format.

Examples

data(mtcars)
mtcars$cyl <- as.factor(mtcars$cyl)
summary_vars <- c("mpg", "hp", "wt")
group_var <- "cyl"
table1_gt <- create_table1(data = mtcars, 
                           summary_vars = summary_vars, 
                           group_var = group_var, 
                           title = "Table 1: Population Characteristics by Cylinder",
                           output_format = "gt")
#> Warning: package ‘gt’ was built under R version 4.3.3
#> Warning: package ‘gtsummary’ was built under R version 4.3.3
print(table1_gt)
#> <div id="tafhkeenhy" style="padding-left:0px;padding-right:0px;padding-top:10px;padding-bottom:10px;overflow-x:auto;overflow-y:auto;width:auto;height:auto;">
#>   <style>#tafhkeenhy table {
#>   font-family: system-ui, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji';
#>   -webkit-font-smoothing: antialiased;
#>   -moz-osx-font-smoothing: grayscale;
#> }
#> 
#> #tafhkeenhy thead, #tafhkeenhy tbody, #tafhkeenhy tfoot, #tafhkeenhy tr, #tafhkeenhy td, #tafhkeenhy th {
#>   border-style: none;
#> }
#> 
#> #tafhkeenhy p {
#>   margin: 0;
#>   padding: 0;
#> }
#> 
#> #tafhkeenhy .gt_table {
#>   display: table;
#>   border-collapse: collapse;
#>   line-height: normal;
#>   margin-left: auto;
#>   margin-right: auto;
#>   color: #333333;
#>   font-size: 16px;
#>   font-weight: normal;
#>   font-style: normal;
#>   background-color: #FFFFFF;
#>   width: auto;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #A8A8A8;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #A8A8A8;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#> }
#> 
#> #tafhkeenhy .gt_caption {
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#> }
#> 
#> #tafhkeenhy .gt_title {
#>   color: #333333;
#>   font-size: 125%;
#>   font-weight: initial;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-color: #FFFFFF;
#>   border-bottom-width: 0;
#> }
#> 
#> #tafhkeenhy .gt_subtitle {
#>   color: #333333;
#>   font-size: 85%;
#>   font-weight: initial;
#>   padding-top: 3px;
#>   padding-bottom: 5px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-color: #FFFFFF;
#>   border-top-width: 0;
#> }
#> 
#> #tafhkeenhy .gt_heading {
#>   background-color: #FFFFFF;
#>   text-align: center;
#>   border-bottom-color: #FFFFFF;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #tafhkeenhy .gt_bottom_border {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #tafhkeenhy .gt_col_headings {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #tafhkeenhy .gt_col_heading {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 6px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   overflow-x: hidden;
#> }
#> 
#> #tafhkeenhy .gt_column_spanner_outer {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: normal;
#>   text-transform: inherit;
#>   padding-top: 0;
#>   padding-bottom: 0;
#>   padding-left: 4px;
#>   padding-right: 4px;
#> }
#> 
#> #tafhkeenhy .gt_column_spanner_outer:first-child {
#>   padding-left: 0;
#> }
#> 
#> #tafhkeenhy .gt_column_spanner_outer:last-child {
#>   padding-right: 0;
#> }
#> 
#> #tafhkeenhy .gt_column_spanner {
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: bottom;
#>   padding-top: 5px;
#>   padding-bottom: 5px;
#>   overflow-x: hidden;
#>   display: inline-block;
#>   width: 100%;
#> }
#> 
#> #tafhkeenhy .gt_spanner_row {
#>   border-bottom-style: hidden;
#> }
#> 
#> #tafhkeenhy .gt_group_heading {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   text-align: left;
#> }
#> 
#> #tafhkeenhy .gt_empty_group_heading {
#>   padding: 0.5px;
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   vertical-align: middle;
#> }
#> 
#> #tafhkeenhy .gt_from_md > :first-child {
#>   margin-top: 0;
#> }
#> 
#> #tafhkeenhy .gt_from_md > :last-child {
#>   margin-bottom: 0;
#> }
#> 
#> #tafhkeenhy .gt_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   margin: 10px;
#>   border-top-style: solid;
#>   border-top-width: 1px;
#>   border-top-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 1px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 1px;
#>   border-right-color: #D3D3D3;
#>   vertical-align: middle;
#>   overflow-x: hidden;
#> }
#> 
#> #tafhkeenhy .gt_stub {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #tafhkeenhy .gt_stub_row_group {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   font-size: 100%;
#>   font-weight: initial;
#>   text-transform: inherit;
#>   border-right-style: solid;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   vertical-align: top;
#> }
#> 
#> #tafhkeenhy .gt_row_group_first td {
#>   border-top-width: 2px;
#> }
#> 
#> #tafhkeenhy .gt_row_group_first th {
#>   border-top-width: 2px;
#> }
#> 
#> #tafhkeenhy .gt_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #tafhkeenhy .gt_first_summary_row {
#>   border-top-style: solid;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #tafhkeenhy .gt_first_summary_row.thick {
#>   border-top-width: 2px;
#> }
#> 
#> #tafhkeenhy .gt_last_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #tafhkeenhy .gt_grand_summary_row {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   text-transform: inherit;
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #tafhkeenhy .gt_first_grand_summary_row {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-top-style: double;
#>   border-top-width: 6px;
#>   border-top-color: #D3D3D3;
#> }
#> 
#> #tafhkeenhy .gt_last_grand_summary_row_top {
#>   padding-top: 8px;
#>   padding-bottom: 8px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#>   border-bottom-style: double;
#>   border-bottom-width: 6px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #tafhkeenhy .gt_striped {
#>   background-color: rgba(128, 128, 128, 0.05);
#> }
#> 
#> #tafhkeenhy .gt_table_body {
#>   border-top-style: solid;
#>   border-top-width: 2px;
#>   border-top-color: #D3D3D3;
#>   border-bottom-style: solid;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#> }
#> 
#> #tafhkeenhy .gt_footnotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #tafhkeenhy .gt_footnote {
#>   margin: 0px;
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #tafhkeenhy .gt_sourcenotes {
#>   color: #333333;
#>   background-color: #FFFFFF;
#>   border-bottom-style: none;
#>   border-bottom-width: 2px;
#>   border-bottom-color: #D3D3D3;
#>   border-left-style: none;
#>   border-left-width: 2px;
#>   border-left-color: #D3D3D3;
#>   border-right-style: none;
#>   border-right-width: 2px;
#>   border-right-color: #D3D3D3;
#> }
#> 
#> #tafhkeenhy .gt_sourcenote {
#>   font-size: 90%;
#>   padding-top: 4px;
#>   padding-bottom: 4px;
#>   padding-left: 5px;
#>   padding-right: 5px;
#> }
#> 
#> #tafhkeenhy .gt_left {
#>   text-align: left;
#> }
#> 
#> #tafhkeenhy .gt_center {
#>   text-align: center;
#> }
#> 
#> #tafhkeenhy .gt_right {
#>   text-align: right;
#>   font-variant-numeric: tabular-nums;
#> }
#> 
#> #tafhkeenhy .gt_font_normal {
#>   font-weight: normal;
#> }
#> 
#> #tafhkeenhy .gt_font_bold {
#>   font-weight: bold;
#> }
#> 
#> #tafhkeenhy .gt_font_italic {
#>   font-style: italic;
#> }
#> 
#> #tafhkeenhy .gt_super {
#>   font-size: 65%;
#> }
#> 
#> #tafhkeenhy .gt_footnote_marks {
#>   font-size: 75%;
#>   vertical-align: 0.4em;
#>   position: initial;
#> }
#> 
#> #tafhkeenhy .gt_asterisk {
#>   font-size: 100%;
#>   vertical-align: 0;
#> }
#> 
#> #tafhkeenhy .gt_indent_1 {
#>   text-indent: 5px;
#> }
#> 
#> #tafhkeenhy .gt_indent_2 {
#>   text-indent: 10px;
#> }
#> 
#> #tafhkeenhy .gt_indent_3 {
#>   text-indent: 15px;
#> }
#> 
#> #tafhkeenhy .gt_indent_4 {
#>   text-indent: 20px;
#> }
#> 
#> #tafhkeenhy .gt_indent_5 {
#>   text-indent: 25px;
#> }
#> 
#> #tafhkeenhy .katex-display {
#>   display: inline-flex !important;
#>   margin-bottom: 0.75em !important;
#> }
#> 
#> #tafhkeenhy div.Reactable > div.rt-table > div.rt-thead > div.rt-tr.rt-tr-group-header > div.rt-th-group:after {
#>   height: 0px !important;
#> }
#> </style>
#>   <table class="gt_table" data-quarto-disable-processing="false" data-quarto-bootstrap="false">
#>   <thead>
#>     <tr class="gt_heading">
#>       <td colspan="5" class="gt_heading gt_title gt_font_normal gt_bottom_border" style>Table 1: Population Characteristics by Cylinder</td>
#>     </tr>
#>     
#>     <tr class="gt_col_headings">
#>       <th class="gt_col_heading gt_columns_bottom_border gt_left" rowspan="1" colspan="1" scope="col" id="label"><span class='gt_from_md'><strong>Variable</strong></span></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="stat_1"><span class='gt_from_md'><strong>4</strong><br />
#> N = 11</span><span class="gt_footnote_marks" style="white-space:nowrap;font-style:italic;font-weight:normal;line-height:0;"><sup>1</sup></span></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="stat_2"><span class='gt_from_md'><strong>6</strong><br />
#> N = 7</span><span class="gt_footnote_marks" style="white-space:nowrap;font-style:italic;font-weight:normal;line-height:0;"><sup>1</sup></span></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="stat_3"><span class='gt_from_md'><strong>8</strong><br />
#> N = 14</span><span class="gt_footnote_marks" style="white-space:nowrap;font-style:italic;font-weight:normal;line-height:0;"><sup>1</sup></span></th>
#>       <th class="gt_col_heading gt_columns_bottom_border gt_center" rowspan="1" colspan="1" scope="col" id="p.value"><span class='gt_from_md'><strong>p-value</strong></span><span class="gt_footnote_marks" style="white-space:nowrap;font-style:italic;font-weight:normal;line-height:0;"><sup>2</sup></span></th>
#>     </tr>
#>   </thead>
#>   <tbody class="gt_table_body">
#>     <tr><td headers="label" class="gt_row gt_left" style="font-weight: bold;">mpg</td>
#> <td headers="stat_1" class="gt_row gt_center">26.0 (22.8, 30.4)</td>
#> <td headers="stat_2" class="gt_row gt_center">19.7 (18.1, 21.0)</td>
#> <td headers="stat_3" class="gt_row gt_center">15.2 (14.3, 16.4)</td>
#> <td headers="p.value" class="gt_row gt_center"><0.001</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left" style="font-weight: bold;">hp</td>
#> <td headers="stat_1" class="gt_row gt_center">91 (65, 97)</td>
#> <td headers="stat_2" class="gt_row gt_center">110 (110, 123)</td>
#> <td headers="stat_3" class="gt_row gt_center">193 (175, 245)</td>
#> <td headers="p.value" class="gt_row gt_center"><0.001</td></tr>
#>     <tr><td headers="label" class="gt_row gt_left" style="font-weight: bold;">wt</td>
#> <td headers="stat_1" class="gt_row gt_center">2.20 (1.84, 2.78)</td>
#> <td headers="stat_2" class="gt_row gt_center">3.22 (2.77, 3.44)</td>
#> <td headers="stat_3" class="gt_row gt_center">3.76 (3.52, 4.07)</td>
#> <td headers="p.value" class="gt_row gt_center"><0.001</td></tr>
#>   </tbody>
#>   
#>   <tfoot class="gt_footnotes">
#>     <tr>
#>       <td class="gt_footnote" colspan="5"><span class="gt_footnote_marks" style="white-space:nowrap;font-style:italic;font-weight:normal;line-height:0;"><sup>1</sup></span> <span class='gt_from_md'>Median (Q1, Q3)</span></td>
#>     </tr>
#>     <tr>
#>       <td class="gt_footnote" colspan="5"><span class="gt_footnote_marks" style="white-space:nowrap;font-style:italic;font-weight:normal;line-height:0;"><sup>2</sup></span> <span class='gt_from_md'>Kruskal-Wallis rank sum test</span></td>
#>     </tr>
#>   </tfoot>
#> </table>
#> </div>

table1_latex <- create_table1(data = mtcars, 
                               summary_vars = summary_vars, 
                               group_var = group_var, 
                               title = "Table 1: Population Characteristics by Cylinder",
                               output_format = "latex")
cat(table1_latex)
#> \begin{landscape}
#> \begin{table}[!t]
#> \fontsize{8.0pt}{10.0pt}\selectfont
#> \begin{tabular*}{\linewidth}{@{\extracolsep{\fill}}lcccc}
#> \toprule
#> \textbf{cyl} & \textbf{4}\\
#> N = 11\textsuperscript{\textit{1}} & \textbf{6}\\
#> N = 7\textsuperscript{\textit{1}} & \textbf{8}\\
#> N = 14\textsuperscript{\textit{1}} & \textbf{p-value}\textsuperscript{\textit{2}} \\ 
#> \midrule\addlinespace[2.5pt]
#> {\bfseries mpg} & 26.0 (22.8, 30.4) & 19.7 (18.1, 21.0) & 15.2 (14.3, 16.4) & <0.001 \\ 
#> {\bfseries hp} & 91 (65, 97) & 110 (110, 123) & 193 (175, 245) & <0.001 \\ 
#> {\bfseries wt} & 2.20 (1.84, 2.78) & 3.22 (2.77, 3.44) & 3.76 (3.52, 4.07) & <0.001 \\ 
#> \bottomrule
#> \end{tabular*}
#> \begin{minipage}{\linewidth}
#> \textsuperscript{\textit{1}}Median (Q1, Q3)\\
#> \textsuperscript{\textit{2}}Kruskal-Wallis rank sum test\\
#> \end{minipage}
#> \end{table}
#> 
#> \end{landscape}